+++ packages/core/action_cable_ext/index.d.ts
@@ -4,6 +4,8 @@ export type ExtendedOptions =
   | Partial<{
       // Initial timestamp to use when requesting stream history during subscription
       historyTimestamp: number | false
+      // Enable PONGs
+      pongs: boolean
     }>
   | Options